refactor: separate multiplexer contracts from backend selection
